dom/workers/test/test_clearTimeouts.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for DOM Worker Threads</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script class="testbody" type="text/javascript">

  new Worker("clearTimeouts_worker.js").onmessage = function(event) {
    event.target.terminate();

    is(event.data, "ready", "Correct message");
    setTimeout(function() { SimpleTest.finish(); }, 1000);
  }

  SimpleTest.waitForExplicitFinish();
  SimpleTest.requestFlakyTimeout("untriaged");

</script>
</pre>
</body>
</html>
